Pull out LocalizedName into its own file so both LanguageName and CountryName can use it
